Fix unittests for backends

+++ tests/tests_catapult.sh
@@ -56,13 +56,17 @@ testConfig() {
 
 # Tests backend switch
 testBackend() {
-  rm -rf buildtest
-  BACKEND=gke make buildir
-  assertTrue 'create buildir' "[ -d 'buildtest' ]"
-  ENVRC="$(cat "$PWD"/buildtest/.envrc)"
-  assertContains 'contains BACKEND' "$ENVRC" 'BACKEND=gke'
-  BACKEND=gke make clean
-  assertTrue 'clean buildir' "[ ! -d 'buildtest' ]"
+    for b in kind caasp4os eks gke imported minikube; do
+        if [ -d "$ROOT_DIR/backend/$b" ]; then
+            rm -rf buildtest
+            BACKEND="$b" make buildir
+            assertTrue 'create buildir' "[ -d 'buildtest' ]"
+            ENVRC="$(cat "$PWD"/buildtest/.envrc)"
+            assertContains 'contains BACKEND' "$ENVRC" BACKEND="$b"
+            BACKEND="$b" make clean
+            assertTrue 'clean buildir' "[ ! -d 'buildtest' ]"
+        fi
+    done
 }
 
 # Tests imported backend
@@ -88,12 +92,12 @@ testJson() {
   unset BACKEND
   rm -rf buildjson
   echo '{ "BACKEND": "gke", "CLUSTER_NAME": "json" }' > test.json
-  CONFIG=$PWD/test.json make buildir
+  BACKEND=gke CONFIG=$PWD/test.json make buildir
   assertTrue 'create buildir' "[ -d 'buildjson' ]"
   ENVRC="$(cat "$PWD"/buildjson/.envrc)"
   assertContains 'contains BACKEND' "$ENVRC" 'BACKEND=gke'
   assertContains 'contains CLUSTER_NAME' "$ENVRC" 'CLUSTER_NAME=json'
-  CONFIG=$PWD/test.json make clean
+  BACKEND=gke CONFIG=$PWD/test.json make clean
   assertTrue 'clean buildir' "[ ! -d 'buildjson' ]"
   rm -rf test.json
 }
